Move TAS lookup code into a more accessible place
We'll be using this for the staging models, too
