+++ dataactcore/models/domainModels.py
@@ -1,6 +1,10 @@
+from datetime import timedelta
+
+import sqlalchemy as sa
 from sqlalchemy import (
     Column, Date, ForeignKey, Index, Integer, Numeric, Text, UniqueConstraint)
 from sqlalchemy.orm import relationship
+
 from dataactcore.models.baseModel import Base
 
 
@@ -32,6 +36,9 @@ TAS_COMPONENTS = (
 
 
 class TASLookup(Base) :
+    """An entry of CARS history -- this TAS was present in the CARS file
+    between internal_start_date and internal_end_date (potentially null)
+    """
     __tablename__ = "tas_lookup"
     tas_id = Column(Integer, primary_key=True)
     account_num = Column(Integer, index=True, nullable=False)
@@ -77,6 +84,40 @@ Index("ix_tas",
       TASLookup.internal_start_date,
       TASLookup.internal_end_date)
 
+
+def is_not_distinct_from(left, right):
+    """Postgres' IS NOT DISTINCT FROM is an equality check that accounts for
+    NULLs. Unfortunately, it doesn't make use of indexes. Instead, we'll
+    imitate it here"""
+    return sa.or_(left == right, sa.and_(left.is_(None), right.is_(None)))
+
+
+def matching_cars_subquery(sess, model_class, start_date, end_date):
+    """We frequently need to mass-update records to look up their CARS history
+    entry. This function creates a subquery to be used in that update call. We
+    pass in the database session to avoid circular dependencies"""
+    # Why min()?
+    # Our data schema doesn't prevent two TAS history entries with the same
+    # TAS components (ATA, AI, etc.) from being valid at the same time. When
+    # that happens (unlikely), we select the minimum (i.e. older) of the
+    # potential TAS history entries.
+    subquery = sess.query(sa.func.min(TASLookup.tas_id))
+
+    # Filter to matching TAS components, accounting for NULLs
+    for field_name in TAS_COMPONENTS:
+        tas_col = getattr(TASLookup, field_name)
+        model_col = getattr(model_class, field_name)
+        subquery = subquery.filter(is_not_distinct_from(tas_col, model_col))
+
+    day_after_end = end_date + timedelta(days=1)
+    model_dates = sa.tuple_(start_date, end_date)
+    tas_dates = sa.tuple_(TASLookup.internal_start_date,
+                          sa.func.coalesce(TASLookup.internal_end_date,
+                                           day_after_end))
+    subquery = subquery.filter(model_dates.op('OVERLAPS')(tas_dates))
+    return subquery.as_scalar()
+
+
 class CGAC(Base):
     __tablename__ = "cgac"
     cgac_id = Column(Integer, primary_key=True)
@@ -90,6 +131,7 @@ class ObjectClass(Base):
     object_class_name = Column(Text)
 
 class SF133(Base):
+    """Represents GTAS records"""
     __tablename__ = "sf_133"
     sf133_id = Column(Integer,primary_key=True)
     agency_identifier = Column(Text, nullable=False)
